+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to mainline dht will be documented in this file.
 
 - Export `errors` module containing `PutError` as a part of the response of `Rpc::put`.
 - `Dht::id()` and `AsyncDht::id()` to get the node's Id.
+- `Dht::find_node()` and `AsyncDht::find_node()` to lookup a certain target, without calling `get_peers` and return a routing table of the responding nodes.
 - `Dht::routing_table()` and `AsyncDht::routing_table()` to get a copy of the routing table.
 - `measure_dht` example to estimate the DHT size.